moved color picker out of column component
