+++ src/components/column/index.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { Droppable } from 'react-beautiful-dnd';
-import Task from './task';
-import boardState from '../util/board-state';
+import Task from '../task';
+import boardState from '../../util/board-state';
 import { ControlledMenu, MenuItem, useMenuState } from '@szhsin/react-menu';
-import { ColumnJson } from '../util/kanban-types';
+import { ColumnJson } from '../../util/kanban-types';
+import ColorPicker from './color-picker';
 
 let IdOfTaskJustAdded = '';
 
@@ -17,7 +18,15 @@ let IdOfTaskJustAdded = '';
  * A StrictColumnJSON passed in will update the data prop of this Column to the parameter. If this Columns' id (a string)
  * is given, then this Column will be deleted.
  */
-function Column({ data, numCols, index }: { data: ColumnJson; numCols: number; index: number }) {
+export default function Column({
+    data,
+    numCols,
+    index,
+}: {
+    data: ColumnJson;
+    numCols: number;
+    index: number;
+}) {
     // open and close context menu
     const { toggleMenu, ...menuProps } = useMenuState();
     const [menuAnchorPoint, setMenuAnchorPoint] = React.useState({
@@ -163,7 +172,7 @@ function Column({ data, numCols, index }: { data: ColumnJson; numCols: number; i
             <ColorPicker
                 isOpen={colorPickerOpen}
                 color={data.color}
-                changeColor={(color) => boardState.setColumnColor(data.id, color)}
+                changeColor={(newColor) => boardState.setColumnColor(data.id, newColor)}
             />
 
             {/* Add Task Button */}
@@ -190,10 +199,7 @@ function Column({ data, numCols, index }: { data: ColumnJson; numCols: number; i
                     <div
                         ref={provided.innerRef}
                         {...provided.droppableProps}
-                        className={[
-                            'column-tasks',
-                            snapshot.isDraggingOver ? 'drag-over' : '',
-                        ].join(' ')}
+                        className={`column-tasks ${snapshot.isDraggingOver ? 'drag-over' : ''}`}
                     >
                         {data.tasks.map((task, taskIndex) => (
                             <Task
@@ -213,87 +219,3 @@ function Column({ data, numCols, index }: { data: ColumnJson; numCols: number; i
         </div>
     );
 }
-
-const darkSwatches = [
-    // colors to pick from in dark mode
-    '#dd302a',
-    '#cf4d19',
-    '#ec9c25',
-    '#7ac41a',
-    '#416a0b',
-    '#338c84',
-    '#344fa2',
-    '#d741e3',
-    '#9900ef',
-    '#6a6a6a',
-];
-const lightSwatches = [
-    // colors to pick from in light mode
-    '#ff6900',
-    '#fcb900',
-    '#7bdcb5',
-    '#00d084',
-    '#8ed1fc',
-    '#0693e3',
-    '#abb8c3',
-    '#eb144c',
-    '#f78da7',
-    '#9900ef',
-];
-
-function ColorPicker({
-    isOpen,
-    color,
-    changeColor,
-}: {
-    isOpen: boolean;
-    color: string;
-    changeColor: (color: string) => void;
-}): JSX.Element {
-    const [textColor, setTextColor] = React.useState(color.slice(1));
-    const swatches = boardState.isLightMode ? lightSwatches : darkSwatches;
-
-    const colorPickerStyle = {
-        // CSS styles so that color picker "swipes" open and closed
-        maxHeight: isOpen ? '6rem' : 0,
-        pointerEvents: isOpen ? 'all' : 'none',
-        transition: 'max-height 0.4s linear',
-    } as const;
-
-    return (
-        <div className="column-color-picker" style={colorPickerStyle}>
-            {swatches.map((swatch) => (
-                <button
-                    key={swatch}
-                    className="column-color-picker__swatch"
-                    style={{ backgroundColor: swatch }}
-                    onClick={() => changeColor(swatch)}
-                />
-            ))}
-            <div className="text-picker">
-                <div className="input-tag"> # </div>
-                <input
-                    value={textColor}
-                    onChange={(e) => {
-                        const newColor = e.target.value;
-                        if (newColor.length > 6 || !/^[\da-fA-F]*$/.test(newColor)) {
-                            return;
-                        }
-                        setTextColor(newColor);
-                    }}
-                    onBlur={() => {
-                        if (textColor.length === 3) {
-                            changeColor(
-                                `#${textColor[0]}${textColor[0]}${textColor[1]}${textColor[1]}${textColor[2]}${textColor[2]}`
-                            );
-                        } else if (textColor.length === 6) {
-                            changeColor(`#${textColor}`);
-                        }
-                    }}
-                />
-            </div>
-        </div>
-    );
-}
-
-export default Column;
